feat: introducing VSR protocol
Initial reconciliation (recovery) implementation done (not tested).

Refs: #109

+++ evita_engine/src/main/java/io/evitadb/spi/cluster/protocol/reconfiguration/ReconfigurationRequest.java
@@ -6,7 +6,7 @@
  *             |  __/\ V /| | || (_| | |_| | |_) |
  *              \___| \_/ |_|\__\__,_|____/|____/
  *
- *   Copyright (c) 2025
+ *   Copyright (c) 2025-2026
  *
  *   Licensed under the Business Source License, Version 1.1 (the "License");
  *   you may not use this file except in compliance with the License.
@@ -67,7 +67,7 @@ import java.net.InetAddress;
  * @see StartEpochRequest
  */
 public record ReconfigurationRequest(
-	int epoch,
+	long epoch,
 	long timestamp,
 	@Nonnull InetAddress[] clusterMembers
 ) implements Serializable {
