+++ evita_engine/src/main/java/io/evitadb/spi/cluster/protocol/viewChange/DoViewChangeResponse.java
@@ -6,7 +6,7 @@
  *             |  __/\ V /| | || (_| | |_| | |_) |
  *              \___| \_/ |_|\__\__,_|____/|____/
  *
- *   Copyright (c) 2025
+ *   Copyright (c) 2025-2026
  *
  *   Licensed under the Business Source License, Version 1.1 (the "License");
  *   you may not use this file except in compliance with the License.
@@ -42,7 +42,7 @@ import io.evitadb.spi.cluster.protocol.HashChainedClusterResponseMessage;
 public record DoViewChangeResponse(
 	int selfIndex,
 	long crc32,
-	int epoch,
+	long epoch,
 	long viewNumber
 ) implements HashChainedClusterResponseMessage {
 }
